requirements.txt
openai
python-dotenv
PyMuPDF
redis
streamlit
